HW13 refactoring

+++ hw12_13_14_15_calendar/internal/storage/sql/storage.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"time"
 
 	"github.com/Remneva/otus_hw/hw12_13_14_15_calendar/internal/storage"
 
@@ -20,8 +21,12 @@ type Storage struct {
 	storage.EventsStorage
 }
 
-func (s *Storage) Connect(ctx context.Context, dsn string, l *zap.Logger) (err error) {
-	s.l = l
+func NewStorage(l *zap.Logger) *Storage {
+	s := &Storage{
+		l: l}
+	return s
+}
+func (s *Storage) Connect(ctx context.Context, dsn string) (err error) {
 	s.db, err = sql.Open("pgx", dsn)
 	if err != nil {
 		s.l.Error("Error", zap.String("Open connection", err.Error()))
@@ -40,7 +45,14 @@ func (s *Storage) Close() error {
 }
 
 func (s *Storage) DeleteEvent(ctx context.Context, id int64) error {
-	_, err := s.db.ExecContext(ctx, `
+	exist, err := s.eventExistValidationByID(id)
+	if err != nil {
+		return fmt.Errorf("SELECT query error %w", err)
+	}
+	if !exist {
+		return fmt.Errorf("event does not exist %w", err)
+	}
+	_, err = s.db.ExecContext(ctx, `
 		DELETE from events where ID = $1
 		`, id)
 	if err != nil {
@@ -52,6 +64,13 @@ func (s *Storage) DeleteEvent(ctx context.Context, id int64) error {
 }
 
 func (s *Storage) UpdateEvent(ctx context.Context, ev storage.Event) error {
+	exist, err := s.eventExistValidationByID(ev.ID)
+	if err != nil {
+		return fmt.Errorf("SELECT query error %w", err)
+	}
+	if !exist {
+		return fmt.Errorf("event does not exist %w", err)
+	}
 	query := "Update events SET owner = $1, title = $2, description = $3, start_date = $4, start_time = $5, end_date = $6, end_time = $7 WHERE id = $8 "
 	result, err := s.db.ExecContext(ctx, query, ev.Owner, ev.Title, ev.Description, ev.StartDate, ev.StartTime, ev.EndDate, ev.EndTime, ev.ID)
 	if err != nil {
@@ -66,6 +85,7 @@ func (s *Storage) UpdateEvent(ctx context.Context, ev storage.Event) error {
 		s.l.Info("Event updated", zap.Int64("id", ev.ID))
 	} else {
 		s.l.Info("Event does not exist", zap.Int64("id", ev.ID))
+		return fmt.Errorf("event does not exist")
 	}
 
 	return nil
@@ -73,9 +93,16 @@ func (s *Storage) UpdateEvent(ctx context.Context, ev storage.Event) error {
 
 func (s *Storage) GetEvent(ctx context.Context, id int64) (storage.Event, error) {
 	var ev storage.Event
+	exist, err := s.eventExistValidationByID(id)
+	if err != nil {
+		return ev, fmt.Errorf("SELECT query error %w", err)
+	}
+	if !exist {
+		return ev, fmt.Errorf("event does not exist %w", err)
+	}
 	row := s.db.QueryRowContext(ctx, `
 		SELECT id, owner, title, description, start_date, start_time, end_date, end_time FROM events where id = $1`, id)
-	err := row.Scan(
+	err = row.Scan(
 		&ev.ID,
 		&ev.Owner,
 		&ev.Title,
@@ -85,7 +112,7 @@ func (s *Storage) GetEvent(ctx context.Context, id int64) (storage.Event, error)
 		&ev.EndDate,
 		&ev.EndTime)
 	if err != nil {
-		return ev, fmt.Errorf("open connection error %w", err)
+		return ev, fmt.Errorf("query error %w", err)
 	}
 	return ev, nil
 }
@@ -120,9 +147,16 @@ func (s *Storage) GetEvents(ctx context.Context) ([]storage.Event, error) {
 }
 
 func (s *Storage) AddEvent(ctx context.Context, ev storage.Event) (int64, error) {
+	exist, err := s.eventExistValidation(ev.Owner, ev.StartTime)
+	if err != nil {
+		return 0, fmt.Errorf("SELECT query error %w", err)
+	}
+	if exist {
+		return 0, fmt.Errorf("event already exist at this time")
+	}
 	query := `INSERT INTO events (owner, title, description, start_date, start_time, end_date, end_time)
 VALUES($1, $2, $3, $4, $5, $6, $7)`
-	_, err := s.db.ExecContext(ctx, query, ev.Owner, ev.Title, ev.Description, ev.StartDate, ev.StartTime, ev.EndDate, ev.EndTime)
+	_, err = s.db.ExecContext(ctx, query, ev.Owner, ev.Title, ev.Description, ev.StartDate, ev.StartTime, ev.EndDate, ev.EndTime)
 	if err != nil {
 		return 0, fmt.Errorf("open connection error %w", err)
 	}
@@ -134,4 +168,24 @@ VALUES($1, $2, $3, $4, $5, $6, $7)`
 	}
 	s.l.Info("Event Created", zap.Int64("id", id))
 	return id, nil
+}
+
+func (s *Storage) eventExistValidationByID(id int64) (bool, error) {
+	var exists bool
+	row := s.db.QueryRow("SELECT EXISTS(SELECT * FROM events WHERE id = $1)", id)
+	if err := row.Scan(&exists); err != nil {
+		s.l.Error("Select query error", zap.Error(err))
+		return exists, fmt.Errorf("query error %w", err)
+	}
+	return exists, nil
+}
+
+func (s *Storage) eventExistValidation(owner int64, time time.Time) (bool, error) {
+	var exists bool
+	row := s.db.QueryRow("SELECT EXISTS(SELECT * FROM events WHERE owner = $1 AND start_time = $2)", owner, time)
+	if err := row.Scan(&exists); err != nil {
+		s.l.Error("Select query error", zap.Error(err))
+		return exists, fmt.Errorf("query error %w", err)
+	}
+	return exists, nil
 }
